services/auth-service/requirements.txt
# WaqediAI Auth Service Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
redis>=5.0.0
httpx>=0.26.0
google-auth>=2.25.0
google-auth-httplib2>=0.2.0
